scripts/requirements.txt
markdown>=3.4.0
pyyaml>=6.0
Pillow>=10.0.0

